Two new dynamic log functions that display the title of the currently focused window I liked the window-title-in-statusbar feature of dwm very much and wanted to have that in XMonad as well.  Somewhere on the net I found some code to put into Config.hs (and sorry, that was last week and I already forgot where I got it from) which I modified and put into the DynamicLog extension.  One can now set the logHook in Config.hs either to dynamicLogWithTitle to get the usual layout description and workspace list plus window title enclosed in angle brackets, or dynamicLogWithTitleColored "white" (or "red" etc.) to have xmonad print out some ^fg() markers for dzen to display the window title in the given color.
Some windows (like terminals or browsers) change their window title from time
to time but xmonad does not recognize this.  So I started learning Haskell to
provide patches for X11-extras and xmonad so that PropertyNotify events are
captured and, if the event notifies about a WM_NAME property change, call the
logHook to update the status bar.

Hope you find this useful,
  Christian