[Issue 343] Slightly more reliable workarounds in RunOrRaisePrompt
runOrRaisePrompt currently works around firefox and thunderbird 2 by
hard coding the classname, but this doesn't always work. This patch
fixes this by hardcoding the program name for pidof instead, and
trying the regular approach before falling back to the workaround.

The alternative fix is to pass in the workaround function, which is a
more reliable solution. Since the only problem apps seem to be mozilla
apps with versions below 3.0, I opted to keep the hackery in hopes
that we can simply remove it once old firefox and thunderbird stop
being common.