Only report windows on the focused workspace as active for EWMH
Currently, when using xinerama and the EwmhLayoutHook, we report all
visible windows as being on the active workspace. This causes panels'
workspace switchers and the like to think windows are shifting to new
workspaces every time we change workspace.

This patch changes the behaviour so that only the focused workspace is
reported, which fixes this behaviour at the cost of panels' window
lists only reporting the focused windows, rather than all visible
windows (which is a little strange, but makes more sense given
xmonad's workspace model.)